attributes in all the concrete classes

+++ studentOrientation/src/studentOrientation/activity/ActivityInterface.java
@@ -4,5 +4,5 @@ public interface ActivityInterface {
     public double getCost();
     public int getDuration();
     public int getEffort();
-    public int getCarbonFoot();
+    public double getCarbonFoot();
 }
